+++ R/backend-hive.R
@@ -111,17 +111,8 @@ sql_table_analyze.sql_dialect_hive <- function(con, table, ...) {
 }
 
 #' @export
-sql_query_set_op.sql_dialect_hive <- function(
-  con,
-  x,
-  y,
-  method,
-  ...,
-  all = FALSE,
-  lvl = 0
-) {
-  check_bool(all)
+sql_query_set_op.sql_dialect_hive <- function(con, x, y, method, ..., lvl = 0) {
   # compared to default method, can't use parentheses
-  method <- paste0(method, if (all) " ALL")
+  method <- sql_set_op_method(con, method)
   sql_glue2(con, "{x}\n{.sql method}\n{y}")
 }