+++ R/lazy-ops.R
@@ -1,5 +1,6 @@
 #' Lazy operations
 #'
+#' @description
 #' This set of S3 classes describe the action of dplyr verbs. These are
 #' currently used for SQL sources to separate the description of operations
 #' in R from their computation in SQL. This API is very new so is likely
@@ -12,66 +13,6 @@
 #' @keywords internal
 #' @name lazy_ops
 NULL
-
-# Base constructors -------------------------------------------------------
-
-#' @export
-#' @rdname lazy_ops
-lazy_base_query <- function(x, vars, class = character(), ...) {
-  check_character(vars)
-
-  lazy_query(
-    query_type = c(paste0("base_", class), "base"),
-    x = x,
-    vars = vars,
-    ...,
-    group_vars = character(),
-    order_vars = NULL,
-    frame = NULL
-  )
-}
-
-lazy_query_remote <- function(x, vars) {
-  lazy_base_query(x, vars, class = "remote")
-}
-
-base_query <- function(from) {
-  check_table_source(from)
-  query("base", from = from)
-}
-
-#' @export
-sql_build.lazy_base_remote_query <- function(op, con, ...) {
-  base_query(op$x)
-}
-
-#' @export
-sql_build.lazy_base_local_query <- function(op, con, ...) {
-  base_query(op$name)
-}
-
-#' @export
-sql_render.base_query <- function(
-  query,
-  con = NULL,
-  ...,
-  sql_options = NULL,
-  subquery = FALSE,
-  lvl = 0
-) {
-  from <- query$from
-  if (subquery || is.sql(from)) {
-    from
-  } else {
-    from <- sql_escape_table_source(con, from)
-    sql_query_select(con, sql("*"), from, lvl = lvl)
-  }
-}
-
-#' @export
-flatten_query.base_query <- function(qry, query_list, con) {
-  query_list
-}
 
 # op_grps -----------------------------------------------------------------
 
@@ -103,7 +44,7 @@ op_sort.tbl_lazy <- function(op) op_sort(op$lazy_query)
 #' @export
 op_sort.lazy_query <- function(op) {
   # Renaming (like for groups) cannot be done because:
-  # * `order_vars` is a list of quosures
+  # * `order_vars` is a list of expressions
   # * variables needed in sorting can be dropped
   op$order_vars
 }