+++ R/optimise-utils.R
@@ -17,14 +17,11 @@ expr_uses_var <- function(x, vars) {
   if (is.sql(x)) {
     return(TRUE)
   }
-  if (is_call(x, "sql")) {
-    return(TRUE)
-  }
-  if (is_quosure(x)) {
-    return(expr_uses_var(quo_get_expr(x), vars))
-  }
   if (is_symbol(x)) {
     return(is_symbol(x, vars))
+  }
+  if (!is_call(x)) {
+    return(FALSE)
   }
 
   any(purrr::map_lgl(as.list(x[-1]), expr_uses_var, vars))
@@ -38,13 +35,6 @@ expr_uses_sql <- function(x) {
   if (is.sql(x)) {
     return(TRUE)
   }
-  if (is_call(x, "sql")) {
-    return(TRUE)
-  }
-  if (is_quosure(x)) {
-    return(expr_uses_sql(quo_get_expr(x)))
-  }
-
   if (!is_call(x)) {
     return(FALSE)
   }