+++ R/translate-sql.R
@@ -11,7 +11,7 @@
 #'
 #' @param ...,dots Expressions to translate. `translate_sql()`
 #'   automatically quotes them for you.  `translate_sql_()` expects
-#'   a list of already quoted objects.
+#'   a list of expression objects.
 #' @param con Database connection used to determine the SQL dialect.
 #' @param vars_group,vars_order,vars_frame Parameters used in the `OVER`
 #'   expression of windowed functions.
@@ -74,7 +74,7 @@ translate_sql <- function(
   window = TRUE
 ) {
   translate_sql_(
-    quos(...),
+    exprs(...),
     con = con,
     vars_group = vars_group,
     vars_order = vars_order,
@@ -127,10 +127,10 @@ translate_sql_ <- function(
 
   variant <- dbplyr_sql_translation(con)
   pieces <- lapply(dots, function(x) {
-    if (is_null(get_expr(x))) {
+    if (is_null(x)) {
       NULL
-    } else if (is_atomic(get_expr(x))) {
-      escape(get_expr(x), con = con)
+    } else if (is_atomic(x) || is.sql(x)) {
+      escape(x, con = con)
     } else {
       mask <- sql_data_mask(x, variant, con = con, window = window)
       escape(eval_tidy(x, mask), con = con)
@@ -152,7 +152,7 @@ sql_data_mask <- function(
   # Default for unknown functions
   unknown <- setdiff(all_calls(expr), names(variant))
   op <- if (strict) missing_op else default_op
-  top_env <- ceply(unknown, op, parent = empty_env(), env = get_env(expr))
+  top_env <- ceply(unknown, op, parent = empty_env())
 
   # Known R -> SQL functions
   special_calls <- copy_env(variant$scalar, parent = top_env)
@@ -179,15 +179,6 @@ sql_data_mask <- function(
         call = call2(call2("::", sym(pkg), sym(name)))
       )
     }
-  }
-
-  special_calls2$sql <- function(...) {
-    dots <- exprs(...)
-
-    env <- get_env(expr)
-    dots <- purrr::map(dots, eval_tidy, env = env)
-
-    exec(sql, !!!dots)
   }
 
   # Existing symbols in expression
@@ -229,15 +220,8 @@ is_infix_user <- function(x) {
   grepl("^%.*%$", x)
 }
 
-default_op <- function(x, env) {
+default_op <- function(x) {
   check_string(x)
-
-  # Check for shiny reactives; these are zero-arg functions
-  # so need special handling to give a useful error
-  obj <- env_get(env, x, default = NULL, inherit = TRUE)
-  if (inherits(obj, "reactive")) {
-    error_embed("a shiny reactive", "foo()")
-  }
 
   if (is_infix_base(x)) {
     sql_infix(x)
@@ -249,7 +233,7 @@ default_op <- function(x, env) {
   }
 }
 
-missing_op <- function(x, env) {
+missing_op <- function(x) {
   force(x)
 
   function(...) {
@@ -264,9 +248,6 @@ missing_op <- function(x, env) {
 }
 
 all_calls <- function(x) {
-  if (is_quosure(x)) {
-    return(all_calls(quo_get_expr(x)))
-  }
   if (!is.call(x)) {
     return(NULL)
   }
@@ -278,9 +259,6 @@ all_calls <- function(x) {
 all_names <- function(x) {
   if (is.name(x)) {
     return(as.character(x))
-  }
-  if (is_quosure(x)) {
-    return(all_names(quo_get_expr(x)))
   }
   if (!is.call(x)) {
     return(NULL)