+++ tests/testthat/test-verb-summarise.R
@@ -167,7 +167,7 @@ test_that("across doesn't select columns from `.by` #1493", {
 
   out <- lf |>
     summarise(
-      across(everything(), ~ sum(..x, na.rm = TRUE)),
+      across(everything(), ~ sum(.x, na.rm = TRUE)),
       .by = g
     )
 
@@ -261,11 +261,10 @@ test_that("can handle rename", {
     out$lazy_query,
     lazy_select_query(
       x = out$lazy_query$x,
-      select = list(ax = sym("ax"), mean_by = quo(mean(by, na.rm = TRUE))),
+      select = list(ax = sym("ax"), mean_by = expr(mean(by, na.rm = TRUE))),
       group_by = syms("ax"),
       select_operation = "summarise",
       group_vars = character()
-    ),
-    ignore_formula_env = TRUE
+    )
   )
 })